database generation
adds database generation, enlargement. This commit lets the web-app generate its own database if it doesnt exist. it also adds primary key restraints to users.id and products.id in the database, it also adds unique constraints to usernames on the database level, however as it stands now. registering a duplicate username will yield a db-error but no front end indication of such. only error in terminal, not fatal to webapp